@@ -18,19 +18,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.4",
-    "@angular/cdk": "13.3.4",
-    "@angular/common": "13.3.4",
-    "@angular/compiler": "13.3.4",
-    "@angular/core": "13.3.4",
-    "@angular/elements": "13.3.4",
-    "@angular/fire": "7.3.0",
-    "@angular/forms": "13.3.4",
-    "@angular/material": "13.3.4",
-    "@angular/platform-browser": "13.3.4",
-    "@angular/platform-browser-dynamic": "13.3.4",
-    "@angular/router": "13.3.4",
-    "@angular/service-worker": "13.3.4",
+    "@angular/animations": "14.0.0",
+    "@angular/cdk": "14.0.1",
+    "@angular/common": "14.0.0",
+    "@angular/compiler": "14.0.0",
+    "@angular/core": "14.0.0",
+    "@angular/elements": "14.0.0",
+    "@angular/fire": "7.4.1",
+    "@angular/forms": "14.0.0",
+    "@angular/material": "14.0.1",
+    "@angular/platform-browser": "14.0.0",
+    "@angular/platform-browser-dynamic": "14.0.0",
+    "@angular/router": "14.0.0",
+    "@angular/service-worker": "14.0.0",
     "@assemblyscript/loader": "^0.20.6",
     "@datorama/akita": "7.1.1",
     "@datorama/akita-ng-router-store": "7.0.0",
@@ -99,32 +99,32 @@
     "codelyzer": "^6.0.2",
     "cross-env": "^7.0.3",
     "cypress": "^8.3.0",
-    "eslint": "8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.4.0",
     "eslint-plugin-cypress": "^2.12.1",
     "firebase-admin": "^10.0.0",
     "html2canvas": "^1.4.1",
-    "jasmine-core": "~3.10.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
+    "jasmine-core": "4.1.1",
+    "jasmine-spec-reporter": "7.0.0",
+    "jest": "28.1.1",
+    "jest-preset-angular": "12.1.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~4.0.0",
+    "karma-jasmine": "5.0.1",
     "karma-jasmine-html-reporter": "~1.7.0",
     "mime-types": "^2.1.34",
-    "ng-packagr": "13.3.1",
-    "nx": "14.1.9",
+    "ng-packagr": "14.0.0",
+    "nx": "14.2.1",
     "postcss": "^8.3.9",
     "postcss-import": "14.1.0",
     "postcss-preset-env": "7.5.0",
     "postcss-url": "10.1.3",
-    "prettier": "2.5.1",
+    "prettier": "2.6.2",
     "resize-observer-polyfill": "^1.5.1",
-    "ts-jest": "27.1.4",
-    "ts-node": "8.10.2",
-    "typescript": "4.6.3",
+    "ts-jest": "28.0.4",
+    "ts-node": "10.8.1",
+    "typescript": "4.7.3",
     "webpack-bundle-analyzer": "^4.5.0"
   },
   "akitaCli": {
